--- h_std.txt
@@ -1,295 +1,0 @@
-58.280533333333324
-58.46506666666666
-58.53079999999999
-58.755999999999986
-58.92099999999999
-58.975199999999994
-59.27473333333332
-59.44566666666667
-59.4418
-59.72726666666667
-59.88313333333334
-59.89780000000001
-60.27673333333334
-60.45306666666667
-60.6214
-61.01293333333333
-61.241066666666654
-61.40899999999999
-61.79413333333332
-62.046466666666646
-62.17019999999998
-62.61166666666665
-62.879133333333314
-63.01699999999998
-63.332999999999984
-63.58019999999999
-63.624799999999986
-63.7954
-63.96600000000001
-63.97320000000001
-64.22240000000002
-64.3786
-64.40200000000003
-64.66060000000002
-64.80920000000002
-64.78380000000001
-65.00106666666667
-65.12333333333335
-65.20400000000001
-65.49933333333334
-65.64766666666668
-65.796
-66.12353333333334
-66.32346666666668
-66.40320000000001
-66.70013333333333
-66.89426666666665
-66.97839999999998
-67.17453333333331
-67.34166666666665
-67.4138
-67.52753333333334
-67.64126666666668
-67.608
-67.69073333333336
-67.77346666666668
-67.72860000000001
-67.84213333333334
-67.90346666666667
-67.89820000000002
-68.01220000000002
-68.06100000000002
-68.10980000000002
-68.25133333333333
-68.31606666666667
-68.38080000000002
-68.56380000000003
-68.67440000000002
-68.73100000000001
-68.9244
-69.0422
-69.0966
-69.2802
-69.3986
-69.45179999999999
-69.54846666666666
-69.64106666666665
-69.65686666666664
-69.6786
-69.70033333333332
-69.65686666666664
-69.66539999999998
-69.67393333333332
-69.68246666666666
-69.68426666666666
-69.68606666666666
-69.68786666666666
-69.66793333333332
-69.68073333333334
-69.68393333333333
-69.7826
-69.81373333333335
-69.84486666666668
-70.02606666666668
-70.12866666666667
-70.19846666666668
-70.29013333333334
-70.477
-70.53706666666666
-70.63406666666666
-70.82146666666667
-70.94106666666669
-71.01873333333334
-71.12206666666668
-71.15380000000002
-71.15660000000001
-71.15520000000001
-71.1538
-71.12620000000001
-71.11473333333332
-71.07906666666665
-71.04339999999999
-70.99853333333333
-70.95366666666666
-70.9216
-70.86693333333335
-70.81226666666669
-70.77433333333336
-70.73640000000003
-70.58466666666668
-70.62760000000002
-70.77026666666669
-70.83813333333333
-70.90599999999999
-71.07366666666667
-71.17253333333332
-71.30839999999999
-71.41286666666669
-71.47893333333336
-71.57060000000001
-71.7172
-71.7518
-71.8336
-71.83373333333334
-71.83386666666667
-71.8336
-71.89346666666667
-71.82633333333334
-71.77060000000002
-71.69193333333335
-71.61326666666669
-71.5144
-71.42639999999999
-71.3384
-71.2504
-71.1712
-71.092
-70.77519999999998
-70.88753333333332
-70.91746666666666
-70.9474
-71.12819999999998
-71.19339999999998
-71.28139999999999
-71.46566666666668
-71.57773333333334
-71.6572
-71.84853333333334
-71.97226666666667
-72.0284
-72.11726666666667
-72.20613333333334
-72.21260000000001
-72.30473333333333
-72.25866666666666
-72.21260000000001
-72.16073333333334
-72.07686666666667
-71.993
-71.89719999999998
-71.8014
-71.7056
-71.59926666666668
-71.49293333333334
-71.34960000000001
-71.28126666666667
-71.21293333333332
-70.93959999999998
-71.09473333333332
-71.16286666666664
-71.231
-71.41433333333333
-71.51926666666667
-71.5878
-71.78246666666665
-71.90073333333332
-71.98299999999999
-72.08286666666665
-72.18273333333332
-72.20559999999999
-72.26226666666668
-72.23393333333334
-72.20560000000002
-72.17506666666668
-72.10433333333334
-72.0336
-71.94533333333334
-71.85706666666668
-71.76880000000001
-71.65506666666667
-71.54133333333334
-71.40840000000001
-71.31620000000001
-71.224
-70.85520000000001
-70.97193333333333
-71.00886666666666
-71.04579999999999
-71.20386666666667
-71.28113333333333
-71.3326
-71.37806666666667
-71.42353333333332
-71.469
-71.59633333333333
-71.53266666666667
-71.469
-71.40393333333333
-71.31826666666666
-71.2326
-71.12886666666667
-71.02513333333333
-70.92139999999999
-70.7892
-70.657
-70.45519999999999
-70.37153333333332
-70.28786666666666
-69.95320000000001
-70.11946666666668
-70.18013333333334
-70.24080000000001
-70.44326666666667
-70.54453333333333
-70.62480000000001
-70.71039999999999
-70.796
-70.8816
-70.98853333333335
-70.93506666666669
-70.8816
-70.87493333333335
-70.76006666666667
-70.6452
-70.50206666666666
-70.34853333333334
-70.195
-70.0228
-69.828
-69.61180000000002
-69.44133333333333
-69.27086666666668
-68.6786
-68.72926666666666
-68.74193333333334
-68.75459999999998
-68.99453333333332
-69.09106666666666
-69.1876
-69.28226666666666
-69.37693333333333
-69.47160000000001
-69.48880000000001
-69.48020000000001
-69.4716
-69.50006666666665
-69.44713333333333
-69.3942
-69.30340000000001
-69.2056
-69.10780000000001
-68.98153333333335
-68.82626666666667
-68.65499999999999
-68.51933333333332
-68.38366666666666
-67.9156
-67.99533333333335
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
-68.01526666666668
