+++ ref/h_std.txt
@@ -1,0 +1,277 @@
+51.585466666666655
+51.56279999999998
+51.544466666666644
+51.50473333333331
+51.46499999999998
+51.43333333333332
+51.38146666666666
+51.276599999999995
+51.242399999999996
+51.20953333333332
+51.13366666666666
+51.1146
+51.09233333333333
+51.050066666666666
+51.03513333333334
+51.01920000000001
+50.981866666666676
+50.97420000000001
+50.966533333333345
+50.93866666666666
+50.93773333333333
+50.95533333333332
+50.959733333333325
+50.96033333333333
+50.957600000000006
+50.954866666666675
+50.98906666666668
+50.96946666666668
+50.94986666666667
+50.95886666666668
+50.925466666666665
+50.892066666666665
+50.857533333333315
+50.848733333333314
+50.80133333333332
+50.7782
+50.71333333333332
+50.62506666666665
+50.58053333333332
+50.51719999999999
+50.40006666666667
+50.354666666666674
+50.3098
+50.221733333333326
+50.1902
+50.16713333333333
+50.144066666666674
+50.132866666666665
+50.14293333333334
+50.153000000000006
+50.27813333333334
+50.32606666666668
+50.374
+50.428200000000004
+50.60453333333333
+50.68926666666666
+50.773999999999994
+50.87206666666666
+50.96033333333333
+51.02106666666668
+51.09873333333334
+51.115600000000015
+51.173000000000016
+51.230133333333335
+51.21006666666666
+51.24146666666667
+51.248266666666666
+51.255066666666664
+51.24146666666667
+51.273599999999995
+51.221933333333325
+51.18906666666666
+51.12473333333333
+51.054
+50.991800000000005
+50.931333333333335
+50.86826666666667
+50.8044
+50.745733333333334
+50.68706666666667
+50.55460000000001
+50.53893333333333
+50.523266666666665
+50.58619999999999
+50.62986666666666
+50.67353333333332
+50.72659999999999
+50.782599999999995
+50.836400000000005
+50.890200000000014
+50.942133333333345
+50.99406666666667
+51.0336
+51.09479999999999
+51.14799999999999
+51.145999999999994
+51.194133333333326
+51.14406666666665
+51.15079999999999
+51.17139999999999
+51.18159999999999
+51.19179999999999
+51.232733333333336
+51.250066666666676
+51.26740000000001
+51.325333333333326
+51.35486666666666
+51.368199999999995
+51.4406
+51.4768
+51.50299999999999
+51.595866666666666
+51.64373333333334
+51.68120000000001
+51.75946666666667
+51.812333333333335
+51.8416
+51.87566666666667
+51.90973333333333
+51.9154
+51.9322
+51.949000000000005
+51.9296
+51.930666666666674
+51.93173333333334
+51.9296
+51.925399999999996
+51.90659999999999
+51.8878
+51.867866666666664
+51.836933333333334
+51.77320000000001
+51.753866666666674
+51.734533333333346
+51.691800000000015
+51.68213333333335
+51.67246666666668
+51.65180000000001
+51.64893333333334
+51.64606666666667
+51.63460000000001
+51.6598
+51.667599999999986
+51.67539999999998
+51.71619999999999
+51.73499999999999
+51.75179999999999
+51.759666666666654
+51.767533333333326
+51.76979999999999
+51.79326666666666
+51.781533333333336
+51.769800000000004
+51.78386666666667
+51.754533333333335
+51.72520000000001
+51.714866666666666
+51.66733333333334
+51.6198
+51.5844
+51.519800000000004
+51.4244
+51.38033333333334
+51.320866666666674
+51.208200000000005
+51.170733333333345
+51.12686666666667
+51.03960000000001
+51.012466666666676
+50.981933333333345
+50.914200000000015
+50.89646666666668
+50.87873333333335
+50.83180000000001
+50.82853333333335
+50.82526666666668
+50.81220000000001
+50.810533333333346
+50.80646666666668
+50.802400000000006
+50.83246666666667
+50.81453333333334
+50.796600000000005
+50.8058
+50.776
+50.7462
+50.73653333333335
+50.69606666666668
+50.6556
+50.62546666666666
+50.572733333333325
+50.5022
+50.462199999999996
+50.411999999999985
+50.313799999999986
+50.28199999999999
+50.2466
+50.172200000000004
+50.15226666666667
+50.13233333333333
+50.081599999999995
+50.075933333333325
+50.07026666666666
+50.047599999999996
+50.07533333333334
+50.08366666666667
+50.092
+50.126400000000004
+50.143800000000006
+50.16120000000001
+50.1674
+50.1736
+50.1724
+50.19186666666666
+50.182133333333326
+50.172399999999996
+50.17453333333333
+50.15286666666666
+50.13119999999999
+50.11626666666666
+50.08153333333333
+50.046800000000005
+50.018800000000006
+49.9722
+49.8884
+49.858999999999995
+49.822399999999995
+49.745400000000004
+49.723400000000005
+49.7014
+49.6556
+49.64493333333333
+49.63426666666666
+49.60379999999999
+49.6054
+49.6058
+49.6062
+49.6128
+49.64206666666667
+49.654333333333334
+49.66753333333333
+49.680733333333336
+49.690533333333335
+49.69626666666666
+49.69339999999998
+49.69053333333332
+49.704133333333324
+49.68793333333332
+49.671733333333314
+49.65106666666665
+49.65066666666666
+49.61986666666666
+49.60786666666667
+49.562
+49.511733333333325
+49.47613333333334
+49.42733333333333
+49.333733333333335
+49.29993333333334
+49.266133333333336
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
+49.202533333333335
