+++ ref/test.txt
@@ -1,0 +1,277 @@
+51.496
+51.645
+51.645
+51.645
+51.726
+51.726
+51.726
+51.461
+51.461
+51.461
+51.256
+51.256
+51.256
+51.156
+51.156
+51.156
+51.049
+51.049
+51.049
+50.948
+50.948
+50.948
+50.948
+50.938
+50.938
+50.922
+50.922
+50.922
+50.917
+50.917
+50.917
+50.934
+50.934
+50.934
+50.934
+51.014
+51.014
+50.897
+50.897
+50.897
+50.628
+50.628
+50.628
+50.416
+50.416
+50.416
+50.416
+50.223
+50.223
+50.041
+50.041
+50.041
+49.947
+49.947
+49.947
+49.947
+49.943
+49.943
+49.943
+50.070
+50.070
+50.374
+50.374
+50.374
+50.760
+50.760
+50.760
+50.760
+51.218
+51.218
+51.218
+51.267
+51.267
+51.235
+51.235
+51.235
+51.235
+51.231
+51.231
+51.231
+50.862
+50.862
+50.862
+50.443
+50.443
+50.302
+50.302
+50.302
+50.302
+50.289
+50.289
+50.289
+50.351
+50.351
+50.351
+50.627
+50.627
+51.098
+51.098
+51.098
+51.098
+51.109
+51.109
+51.109
+51.068
+51.068
+51.068
+51.149
+51.149
+51.149
+51.199
+51.199
+51.199
+51.251
+51.251
+51.251
+51.369
+51.369
+51.369
+51.511
+51.511
+51.511
+51.692
+51.692
+51.692
+51.917
+51.917
+51.917
+52.044
+52.044
+52.044
+51.880
+51.880
+51.880
+51.763
+51.763
+51.763
+51.708
+51.708
+51.708
+51.635
+51.635
+51.635
+51.580
+51.580
+51.580
+51.590
+51.590
+51.590
+51.618
+51.618
+51.618
+51.665
+51.665
+51.665
+51.752
+51.752
+51.752
+51.862
+51.862
+51.862
+51.708
+51.708
+51.708
+51.442
+51.442
+51.442
+51.225
+51.225
+51.225
+51.039
+51.039
+51.039
+50.893
+50.893
+50.893
+50.816
+50.816
+50.816
+50.784
+50.784
+50.784
+50.767
+50.767
+50.767
+50.773
+50.773
+50.773
+50.844
+50.844
+50.844
+50.755
+50.755
+50.755
+50.515
+50.515
+50.515
+50.320
+50.320
+50.320
+50.166
+50.166
+50.166
+50.053
+50.053
+50.053
+50.002
+50.002
+50.002
+49.984
+49.984
+49.984
+50.021
+50.021
+50.021
+50.081
+50.081
+50.081
+50.178
+50.178
+50.178
+50.263
+50.263
+50.263
+50.077
+50.077
+50.077
+49.875
+49.875
+49.875
+49.756
+49.756
+49.756
+49.657
+49.657
+49.657
+49.564
+49.564
+49.564
+49.528
+49.528
+49.528
+49.545
+49.545
+49.545
+49.596
+49.596
+49.596
+49.663
+49.663
+49.663
+49.663
+49.748
+49.748
+49.726
+49.726
+49.726
+49.502
+49.502
+49.502
+49.353
+49.353
+49.353
+49.353
+49.201
+49.201
+49.060
+49.060
+49.060
+48.994
+48.994
+48.994
+48.995
+48.995
+48.995
+48.995
+48.999
