--- requirements.bat
@@ -1,14 +1,0 @@
-@echo off
-color A
-echo Please wait until installment is completed
-echo installing PyQt5...
-pip install PyQt5
-echo installing opencv-python...
-pip install opencv-python
-echo installing tqdm...
-pip install tqdm
-echo installing rich...
-pip install rich
-echo installing matplotlib...
-pip install matplotlib
-exit
