--- test.txt
@@ -1,295 +1,0 @@
-56.244
-56.244
-56.838
-56.838
-56.838
-57.392
-57.392
-57.392
-58.266
-58.266
-58.266
-58.972
-58.972
-58.972
-59.012
-59.012
-59.012
-59.313
-59.313
-59.313
-59.956
-59.956
-59.956
-60.604
-60.604
-60.604
-61.617
-61.617
-61.617
-62.434
-62.434
-62.434
-63.098
-63.098
-63.098
-63.968
-63.968
-63.968
-64.312
-64.312
-64.312
-64.176
-64.176
-64.176
-64.777
-64.777
-64.777
-65.327
-65.327
-65.327
-65.802
-65.802
-65.802
-66.537
-66.537
-66.537
-67.175
-67.175
-67.175
-67.689
-67.689
-67.689
-67.834
-67.834
-67.834
-67.508
-67.508
-67.508
-67.778
-67.778
-67.778
-68.095
-68.095
-68.095
-68.421
-68.421
-68.421
-68.805
-68.805
-68.805
-69.167
-69.167
-69.167
-69.545
-69.545
-69.545
-69.871
-69.871
-69.871
-69.871
-69.810
-69.810
-69.131
-69.131
-69.131
-69.295
-69.295
-69.295
-69.572
-69.572
-69.572
-69.572
-69.919
-69.919
-70.277
-70.277
-70.277
-70.670
-70.670
-70.670
-70.670
-71.027
-71.027
-71.027
-71.366
-71.366
-71.469
-71.469
-71.469
-70.256
-70.256
-70.256
-70.256
-70.135
-70.135
-70.135
-70.354
-70.354
-70.546
-70.546
-70.546
-70.900
-70.900
-70.900
-70.900
-71.274
-71.274
-71.274
-71.618
-71.618
-71.921
-71.921
-71.921
-71.921
-72.127
-72.127
-72.127
-70.902
-70.902
-70.902
-70.267
-70.267
-70.438
-70.438
-70.438
-70.438
-70.601
-70.601
-70.601
-70.939
-70.939
-70.939
-71.351
-71.351
-71.351
-71.758
-71.758
-71.758
-72.119
-72.119
-72.119
-72.457
-72.457
-72.457
-72.272
-72.272
-72.272
-70.660
-70.660
-70.660
-70.500
-70.500
-70.500
-70.682
-70.682
-70.682
-70.862
-70.862
-70.862
-71.247
-71.247
-71.247
-71.682
-71.682
-71.682
-72.074
-72.074
-72.074
-72.456
-72.456
-72.456
-72.360
-72.360
-72.360
-70.822
-70.822
-70.822
-70.621
-70.621
-70.621
-70.750
-70.750
-70.750
-70.750
-70.750
-70.750
-70.977
-70.977
-70.977
-71.376
-71.376
-71.376
-71.780
-71.780
-71.780
-71.432
-71.432
-71.432
-69.795
-69.795
-69.795
-69.692
-69.692
-69.692
-69.820
-69.820
-69.820
-69.797
-69.797
-69.797
-70.177
-70.177
-70.177
-70.705
-70.705
-70.705
-71.211
-71.211
-71.211
-71.104
-71.104
-71.104
-68.995
-68.995
-68.995
-68.454
-68.454
-68.454
-68.402
-68.402
-68.402
-68.289
-68.289
-68.289
-68.547
-68.547
-68.547
-69.185
-69.185
-69.185
-69.902
-69.902
-69.902
-69.822
-69.822
-69.822
-68.160
-68.160
-68.160
-67.753
-67.753
-67.753
-67.718
-67.718
-67.718
-67.573
-67.573
-67.573
-67.787
-67.787
-67.787
-68.459
-68.459
-68.459
-69.225
