+++ v2.0.ui
@@ -2,34 +2,26 @@
 <ui version="4.0">
  <class>Form</class>
  <widget class="QWidget" name="Form">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>400</width>
-    <height>300</height>
-   </rect>
-  </property>
   <property name="windowTitle">
    <string>Form</string>
   </property>
   <widget class="QLCDNumber" name="lcdNumber">
    <property name="geometry">
     <rect>
-     <x>300</x>
-     <y>270</y>
-     <width>64</width>
-     <height>23</height>
+     <x>493</x>
+     <y>332</y>
+     <width>101</width>
+     <height>61</height>
     </rect>
    </property>
   </widget>
   <widget class="QLabel" name="label">
    <property name="geometry">
     <rect>
-     <x>140</x>
-     <y>270</y>
-     <width>111</width>
-     <height>20</height>
+     <x>10</x>
+     <y>339</y>
+     <width>471</width>
+     <height>51</height>
     </rect>
    </property>
    <property name="text">
@@ -50,8 +42,8 @@
     <rect>
      <x>0</x>
      <y>0</y>
-     <width>401</width>
-     <height>261</height>
+     <width>601</width>
+     <height>331</height>
     </rect>
    </property>
   </widget>
@@ -59,8 +51,8 @@
    <property name="geometry">
     <rect>
      <x>0</x>
-     <y>220</y>
-     <width>401</width>
+     <y>290</y>
+     <width>601</width>
      <height>41</height>
     </rect>
    </property>
